+++ lib/tus/storage/s3.rb
@@ -93,6 +93,7 @@ module Tus
         info.delete("multipart_id")
         info.delete("multipart_parts")
 
+        # Tus server requires us to return the size of the concatenated file.
         client.head_object(bucket: bucket.name, key: object(uid).key).content_length
       rescue
         abort_multipart_upload(multipart_upload) if multipart_upload